pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crop-by-ar"
version = "0.1.0"
description = "A command-line tool to crop images to a specified aspect ratio."
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
readme = "README.md"
requires-python = ">=3.6"
dependencies = [
    "Pillow"
]

[tool.setuptools]
packages = ["src"]

[project.scripts]
crop-by-ar = "src.crop_by_ar:main"
